Only fetch the media files in the trasaction, this way exceptions are thrown when doing other database operations.
